+++ src/lib/utils/date.js
@@ -8,3 +8,13 @@ export const isValid = (date) => {
 
   return true;
 };
+
+
+// returns a date in yyyy-mm-dd format
+export const formattedDate = (date) => {
+  if (isValid(date)) {
+    return new Date(date).toISOString().slice(0, 10);
+  }
+
+  return null;
+};